+++ src/components/CalendarBadge.tsx
@@ -12,7 +12,7 @@ import {
   CALENDAR_BADGE_LABEL,
   CalendarBadgeStatus,
 } from '@/constants/calendar-badge';
-import { CalendarBadgeStyle } from '@/constants/calendar-badge.cva';
+import { CalendarBadgeStyle } from '@/styles/cva/calendar-badge.cva';
 
 interface CalendarBadgeProps {
   status: CalendarBadgeStatus;
